requirements.txt
fastapi
fire
beautifulsoup4
uvicorn
requests-html
webdriver-manager
selenium
sqlitedict